requirements.txt
djangorestframework
markdown
django-filter
whitenoise
mysqlclient
googlemaps
gmplot
django-phonenumber-field
